backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.1.1

# HTTP客户端
aiohttp==3.12.15
httpx==0.25.2

# 数据处理
pydantic==2.5.0
python-multipart==0.0.6

# 音频处理
torch==2.1.0
torchaudio==2.1.0
librosa==0.10.1
soundfile==0.12.1
numpy==1.26.0

# 科学计算
scipy==1.13.1

# 其他工具
loguru==0.7.3
tqdm==4.67.1
